move memento to module with WeightData
